modularizing init.lua: moved lsp and cmp to own files
currently fold.lua is irrelevant. we can delete this later.

LuaSnip is currently for testing. I'm going to make some of my own snippets in this folder, but for now its contents are meaningless.
We source all snippets in the LuaSnip folder in completion.lua

Currently lsp.lua and completion.lua are modules with multiple plugins; I think this makes sense to do because many things go into an lsp and auto-completion.
In the future I may change my mind.