Changes to vimtex and texlab (lsp). Disable compile on save.
Do this because the minted package creates extra files that vimtex/latexmk struggle to keep track of.
This means to compile I need to < lC> and < ll> on occasion after every save.
Interestingly, if I run latexmk <filename> from command line, I get no problems.
Claude says this is because vimtex calls latexmk with the -pvc flag on by default.